New types for script witnesses within the tx body
New type ScriptWitness and Witness. The ScriptWitness is used to
annotate a script into the use site. The Witness type wraps the
ScriptWitness type, with KeyWitness as the other alternative. Which ones
are avilable depends on the witness context. The KeyWitnessInCtx and
ScriptWitnessInCtx types provide the evidence about which kinds of
witness are permitted in each context. In particular the minting context
does not permit key witnesses, only script witnesses.

This is the first step in a refactoring that we will need for Alonzo
support. The refactoring will put the scripts within the tx body at the
sites where they are used. For Shelley era scripts we can do it either
way round, but for Alonzo it makes much more sense to place the script
at the use sites. So to break the Alonzo changes into steps we do the
refactoring for the existing Shelley-based eras first.

The ideas here have been co-authored with Jordan Millar and most of the
code changes here appeared first in his Alonzo branch.

Co-authored-by: Jordan Millar <jordan.millar@iohk.io>